requirements.txt
python-telegram-bot
notion-client
google-generativeai
python-dotenv
